refactor: update token structure and enhance username retrieval in authentication flow
